Added: End of Race Dialog, Extra Customisation to Shell, Patch Extension

+++ Riders.Tweakbox/Components/Netplay/Components/Game/SetupRace.cs
@@ -15,11 +15,11 @@ namespace Riders.Tweakbox.Components.Netplay.Components.Game
 {
     public unsafe class SetupRace : INetplayComponent
     {
-        private static Patch _disableGrandPrixOverridePlayerCount = new Patch((IntPtr)0x0050BC34, new byte[] { 0x90, 0x90, 0x90, 0x90, 0x90, 0x90 });
+        private static Patch _disableGrandPrixOverridePlayerCount = new Patch(0x0050BC34, new byte[] { 0x90, 0x90, 0x90, 0x90, 0x90, 0x90 });
 
         // TODO: Add this in the future without breaking culling.
-        private static Patch _dontSetPlayerTypeOnInit = new Patch((IntPtr) 0x00413E60, new byte[] { 0xEB, 0x27 });
-        private static Patch _dontSetPlayerTypeOnCharSelectInit = new Patch((IntPtr) 0x00463A97, new byte[] { 0xE9, 0xDA, 0x00, 0x00, 0x00 });
+        private static Patch _dontSetPlayerTypeOnInit = new Patch(0x00413E60, new byte[] { 0xEB, 0x27 });
+        private static Patch _dontSetPlayerTypeOnCharSelectInit = new Patch(0x00463A97, new byte[] { 0xE9, 0xDA, 0x00, 0x00, 0x00 });
 
         /// <inheritdoc />
         public Socket Socket { get; set; }
